refactor: change the api directory to differentiate between openfairdb url and next
