+++ server.build.dev.js
@@ -5,7 +5,7 @@ dotenv.config()
 
 process.env.NEXT_PUBLIC_BASICS_API = process.env.NEXT_PUBLIC_BASICS_API || 'https://dev.ofdb.io/v0'
 process.env.NEXT_PUBLIC_SELF_DOMAIN = process.env.NEXT_PUBLIC_SELF_DOMAIN || `https://dev.kartevonmorgen.org`
-process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/v0`
+process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/next/v0`
 process.env.DB_NAME = process.env.DB_NAME || 'kartevonmorgen.development.sqlite'
