+++ server.dev.js
@@ -6,7 +6,7 @@ dotenv.config()
 process.env.HOSTNAME = process.env.HOSTNAME || 'localhost'
 process.env.PORT = process.env.PORT || '3000'
 process.env.NEXT_PUBLIC_SELF_DOMAIN = process.env.NEXT_PUBLIC_SELF_DOMAIN || `http://${process.env.HOSTNAME}:${process.env.PORT}`
-process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/v0`
+process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/next/v0`
 
 
 cli.nextDev(
