+++ server.start.production.js
@@ -5,7 +5,7 @@ process.env.HOSTNAME = process.env.HOSTNAME || 'localhost'
 process.env.PORT = process.env.PORT || '3000'
 process.env.DB_NAME = process.env.DB_NAME || 'kartevonmorgen.sqlite'
 process.env.NEXT_PUBLIC_SELF_DOMAIN = process.env.NEXT_PUBLIC_SELF_DOMAIN || `https://ssr.kartevonmorgen.org`
-process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/v0`
+process.env.NEXT_PUBLIC_SELF_API = process.env.NEXT_PUBLIC_SELF_API || `${process.env.NEXT_PUBLIC_SELF_DOMAIN}/api/next/v0`
 
 
 cli.nextStart(
